server/app/pyproject.toml
[project]
name = "food-dist-server"
version = "0.1.0"
description = "django rest framework server for a food distribution app"
authors = [
    {name = "Kevin Lai"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "django (>=5.1.4,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-stubs (>=5.1.1,<6.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "django-filter (>=24.3,<25.0)",
    "django-phonenumber-field (>=8.0.0,<9.0.0)",
    "phonenumbers (>=8.13.53,<9.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
